Move table update to its own file.
