+++ db/tables/ddl/__init__.py
@@ -1,2 +1,3 @@
 from db.tables.ddl.create import *  # noqa
 from db.tables.ddl.drop import *  # noqa
+from db.tables.ddl.update import *  # noqa
